+++ packages/implementations/react/src/composite/Datepicker/src/components/Footer/footer.types.ts
@@ -1,0 +1,8 @@
+import { ButtonHTMLAttributes } from "react";
+
+export interface FooterProperties {
+  setSelected?: () => void;
+}
+
+export type FooterProps = FooterProperties &
+  ButtonHTMLAttributes<HTMLButtonElement>;