+++ packages/implementations/react/src/composite/Datepicker/src/components/Header/header.types.ts
@@ -1,0 +1,9 @@
+import { HTMLAttributes } from "react";
+import {
+  ReactDatePickerCustomHeaderProps,
+  ReactDatePickerProps,
+} from "react-datepicker";
+
+export type HeaderProps = ReactDatePickerCustomHeaderProps &
+  Pick<ReactDatePickerProps, "showYearPicker" | "showMonthYearPicker"> &
+  HTMLAttributes<HTMLDivElement>;