+++ packages/implementations/react/CHANGELOG.md
@@ -2,12 +2,17 @@
 
 `@rarui-react/components` components is a component library built with [React](https://reactjs.org).
 
+## 2024-06-06 `1.25.0`
+
+#### 🎉 New features
+
+- Added new composite `Datepicker` component. ([#50](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/50) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-06-02 `1.24.1`
 
 #### 💡 Others
 
 - Adjusted texts and description of properties for each typing of all components. ([#46](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/46) by [@jefferson.lima](https://git.rarolabs.com.br/jefferson.lima))
-  yarn
 
 ## 2024-06-02 `1.24.0`